Alert

+++ components/FavoriteComponent.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { View, Flatlist, Text } from 'react-native';
+import { View, Flatlist, Text, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { FlatList } from 'react-native-gesture-handler';
 import { connect } from 'react-redux';
@@ -35,7 +35,26 @@ class Favorites extends Component {
                 {
                     text: 'Delete',
                     type: 'delete',
-                    onPress: () => this.props.deleteFavorite(item.id)
+                    onPress: () => {
+                        Alert.alert(
+                            'Delete Favorite?',
+                            'Are you sure you wish to delete the favorite dish ' + item.name + '?',
+                            [
+                                { 
+                                    text: 'Cancel', 
+                                    onPress: () => console.log(item.name + ' not deleted.'),
+                                    style: 'cancel' 
+                                },
+                                { 
+                                    text: 'OK', 
+                                    onPress: () => this.props.deleteFavorite(item.id),
+                                }
+                            ],
+                            { cancelable: false }
+                        );
+                    } 
+                    
+                    
                 }
             ];
